Consolidate /deposit /withdraw /transfer into /bank
Folds three separate slash commands into one /bank command with
deposit/withdraw/transfer subcommands. Reduces global command count by
2 and groups related actions under a single discoverable namespace.

Also drops the mongoose session transaction from /bank transfer (was
/transfer); standalone MongoDB doesn't support multi-doc transactions
and the prior code would catch the error and report "Failed to
transfer coins" on every attempt. Replaced with an atomic conditional
$inc debit that won't take coins unless the sender has the balance.
