+++ src/commands/utility/help.js
@@ -22,9 +22,7 @@ const CATEGORIES = [
             { name: 'balance',        description: 'Check your wallet and bank balance' },
             { name: 'daily',          description: 'Claim your daily coin reward (24h cooldown)' },
             { name: 'work',           description: 'Earn coins by working a shift (1h cooldown)' },
-            { name: 'deposit',        description: 'Deposit coins into your bank' },
-            { name: 'withdraw',       description: 'Withdraw coins from your bank' },
-            { name: 'transfer',       description: 'Transfer coins to another user' },
+            { name: 'bank',           description: 'Bank actions: deposit, withdraw, transfer' },
             { name: 'inventory',      description: 'View your or another user\'s inventory' },
             { name: 'shop',           description: 'Browse the item shop' },
             { name: 'use',            description: 'Use an item from your inventory' },
